+++ wb/wb.h
@@ -324,8 +324,8 @@ typedef struct
 typedef struct _wbo
 {							// wbo
 	HWND hwnd;				// Control or window handle
-	UINT id;				// Control identifier
-	UINT uClass;			// Object class
+	UINT64 id;				// Control identifier
+	UINT64 uClass;			// Object class
 	int item;				// Item index
 	int subitem;			// Sub-item index
 	DWORD style;			// WinBinder style
@@ -356,8 +356,8 @@ typedef struct
 
 typedef struct
 { // td
-	UINT nCtrls;
-	UINT nPages;
+	UINT64 nCtrls;
+	UINT64 nPages;
 	struct
 	{
 		HWND hwnd;
@@ -365,7 +365,7 @@ typedef struct
 	struct
 	{
 		HWND hwnd;
-		UINT id;
+		UINT64 id;
 		BYTE nTab;
 	} ctrl[MAX_TABCONTROLS];
 } TABDATA, *PTABDATA;
@@ -440,14 +440,14 @@ HDC wbCreateBitmapDC(HBITMAP hbm);
 BOOL wbCopyBits(HDC hdc, HBITMAP hBitmap, int xTarget, int yTarget);
 BOOL wbCopyPartialBits(HDC hdc, HBITMAP hBitmap, int xTarget, int yTarget, int nWidth, int nHeight, int xSource, int ySource);
 BOOL wbMaskPartialBits(HDC hdc, HBITMAP hbmImage, HBITMAP hbmMask, int xTarget, int yTarget, int nWidth, int nHeight, int xSource, int ySource);
-//HANDLE		wbLoadImage(LPCTSTR pszImageFile, UINT nIndex);
-HANDLE wbLoadImage(LPCTSTR pszImageFile, UINT nIndex, LPARAM lParam);
+//HANDLE		wbLoadImage(LPCTSTR pszImageFile, UINT64 nIndex);
+HANDLE wbLoadImage(LPCTSTR pszImageFile, UINT64 nIndex, LPARAM lParam);
 DWORD wbGetBitmapBits(HBITMAP hbm, BYTE **lpBits, BOOL bCompress4to3);
 COLORREF wbGetPixelDirect(unsigned char *pixdata, int xPos, int yPos, BOOL bCompress4to3);
 
 // WB_CONTROL.C
 
-PWBOBJ wbCreateControl(PWBOBJ pwboParent, UINT uWinBinderClass, LPCTSTR pszCaption, LPCTSTR pszTooltip, int xPos, int yPos, int nWidth, int nHeight, UINT64 id, DWORD dwWinBinderStyle, LONG_PTR lParam, int nTab);
+PWBOBJ wbCreateControl(PWBOBJ pwboParent, UINT64 uWinBinderClass, LPCTSTR pszCaption, LPCTSTR pszTooltip, int xPos, int yPos, int nWidth, int nHeight, UINT64 id, DWORD dwWinBinderStyle, LONG_PTR lParam, int nTab);
 BOOL wbDestroyControl(PWBOBJ pwbo);
 PWBOBJ wbGetControl(PWBOBJ pwboParent, int id);
 BOOL wbCreateItem(PWBOBJ pwbo, LPCTSTR pszItemText);
@@ -461,13 +461,13 @@ BOOL wbSetRange(PWBOBJ pwbo, LONG_PTR dwMin, LONG_PTR dwMax);
 BOOL wbGetVisible(PWBOBJ pwbo);
 BOOL wbSetVisible(PWBOBJ pwbo, BOOL bState);
 BOOL wbSetStyle(PWBOBJ pwbo, DWORD dwWBStyle, BOOL bSet);
-BOOL wbIsValidClass(UINT uClass);
+BOOL wbIsValidClass(UINT64 uClass);
 BOOL wbGetEnabled(PWBOBJ pwbo);
 BOOL wbSetEnabled(PWBOBJ pwbo, BOOL bState);
 BOOL wbSetText(PWBOBJ pwbo, LPCTSTR pszText, int nItem, BOOL bTooltip);
-BOOL wbGetText(PWBOBJ pwbo, LPTSTR pszText, UINT nMaxChars, int nIndex);
+BOOL wbGetText(PWBOBJ pwbo, LPTSTR pszText, UINT64 nMaxChars, int nIndex);
 BOOL wbGetRtfText(PWBOBJ pwbo, char **unc);
-UINT wbGetTextLength(PWBOBJ pwbo, int nIndex);
+UINT64 wbGetTextLength(PWBOBJ pwbo, int nIndex);
 BOOL wbDeleteItems(PWBOBJ pwbo, BOOL bClearAll);
 BOOL wbRefreshControl(PWBOBJ pwbo, int xpos, int ypos, int nWidth, int nHeight, BOOL bNow);
 DWORD wbGetSelected(PWBOBJ pwbo);
@@ -539,16 +539,16 @@ BOOL wbSetTabControlItemImages(PWBOBJ pwbo, int item, int nImageIndex);
 
 COLORREF wbGetPixel(HANDLE hImage, int xPos, int yPos);
 BOOL wbSetPixel(HANDLE hImage, int xPos, int yPos, COLORREF color);
-BOOL wbDrawLine(HANDLE handle, int x0, int y0, int x1, int y1, COLORREF cl, UINT nLineWidth, UINT nLineStyle);
-BOOL wbDrawRect(HANDLE handle, int xPos, int yPos, int nWidth, int nHeight, COLORREF cl, BOOL bFilled, UINT nLineWidth, UINT nLineStyle);
-BOOL wbDrawEllipse(HANDLE handle, int xPos, int yPos, int nWidth, int nHeight, COLORREF cl, BOOL bFilled, UINT nLineWidth, UINT nLineStyle);
+BOOL wbDrawLine(HANDLE handle, int x0, int y0, int x1, int y1, COLORREF cl, UINT64 nLineWidth, UINT64 nLineStyle);
+BOOL wbDrawRect(HANDLE handle, int xPos, int yPos, int nWidth, int nHeight, COLORREF cl, BOOL bFilled, UINT64 nLineWidth, UINT64 nLineStyle);
+BOOL wbDrawEllipse(HANDLE handle, int xPos, int yPos, int nWidth, int nHeight, COLORREF cl, BOOL bFilled, UINT64 nLineWidth, UINT64 nLineStyle);
 BOOL wbDrawText(HANDLE handle, LPCTSTR pszString, int xPos, int yPos, int nWidth, int nHeight, int nFont, DWORD dwFlags);
 BOOL wbDrawBitmap(HANDLE handle, HBITMAP hbmBits, int xPos, int yPos, int nWidth, int nHeight, int xOffset, int yOffset, COLORREF clTransparent);
 BOOL wbGetTextSize(PSIZE psizeText, LPCTSTR pszString, int nFont);
 
 // WB_LOWLEVEL.C
 
-LPARAM wbSendMessage(PWBOBJ pwbo, UINT uMsg, WPARAM wParam, LPARAM lParam);
+LPARAM wbSendMessage(PWBOBJ pwbo, UINT64 uMsg, WPARAM wParam, LPARAM lParam);
 HMODULE wbLoadLibrary(LPCTSTR pszLibName);
 FARPROC wbGetLibraryFunction(HMODULE hLib, LPCSTR pszFunction);
 BOOL wbReleaseLibrary(HMODULE hLib);
@@ -577,20 +577,20 @@ PFONT wbSysDlgFont(PWBOBJ pwboParent, LPCTSTR pszTitle, PFONT pfont);
 
 // WB_WINDOW.C
 
-PWBOBJ wbCreateWindow(PWBOBJ pwboParent, UINT uWinBinderClass, LPCTSTR pszCaption, LPCTSTR pszTooltip, int xPos, int yPos, int nWidth, int nHeight, UINT id, DWORD dwWBStyle, LONG_PTR lParam);
+PWBOBJ wbCreateWindow(PWBOBJ pwboParent, UINT64 uWinBinderClass, LPCTSTR pszCaption, LPCTSTR pszTooltip, int xPos, int yPos, int nWidth, int nHeight, UINT64 id, DWORD dwWBStyle, LONG_PTR lParam);
 BOOL wbDestroyWindow(PWBOBJ pwbo);
 BOOL wbSetWindowHandler(PWBOBJ pwbo, LPDWORD pszObj, LPCTSTR pszHandler);
 BOOL wbSetWindowSize(PWBOBJ pwbo, int nWidth, int nHeight, int nShowMode);
 DWORD wbGetWindowSize(PWBOBJ pwbo, BOOL bClientRect);
 DWORD wbGetWindowPosition(PWBOBJ pwbo, PWBOBJ pwboParent, BOOL bClientRect);
 BOOL wbSetWindowPosition(PWBOBJ pwbo, int xPos, int yPos, PWBOBJ pwboParent);
-// void *wbGetAppInfo(LPCTSTR pszInfo, BOOL *pbString, LPTSTR pszString, UINT uLen);
+// void *wbGetAppInfo(LPCTSTR pszInfo, BOOL *pbString, LPTSTR pszString, UINT64 uLen);
 // BOOL wbSetAppInfo(LPCTSTR pszInfo, BOOL *pbString, DWORD dwValue);
-BOOL wbSetTimer(PWBOBJ pwbo, int id, UINT uPeriod);
+BOOL wbSetTimer(PWBOBJ pwbo, int id, UINT64 uPeriod);
 BOOL wbSortLVColumn(PWBOBJ pwbo, int nSubItem, BOOL bAscending);
-UINT wbGetControlList(PWBOBJ pwboParent, PWBOBJ pwboList[], UINT nMaxControls);
+UINT64 wbGetControlList(PWBOBJ pwboParent, PWBOBJ pwboList[], UINT64 nMaxControls);
 HWND wbGetRequestedAppWindow(LPCTSTR pszCaption, BOOL bBringToFront);
-BOOL wbSetWindowArea(PWBOBJ pwbo, UINT type, int xPos, int yPos, int nWidth, int nHeight);
+BOOL wbSetWindowArea(PWBOBJ pwbo, UINT64 type, int xPos, int yPos, int nWidth, int nHeight);
 
 // WB_WINSYS.C
 
@@ -605,20 +605,20 @@ BOOL wbPlaySystemSound(int nStyle);
 BOOL wbPlaySound(LPCTSTR pszFileName, LPCTSTR pszCommand);
 BOOL wbStopSound(LPCTSTR pszCommand);
 BOOL wbShowLastError(LPCTSTR pszCaption, BOOL bMessageBox);
-int wbMessageBox(PWBOBJ pwboParent, LPCTSTR pszText, LPCTSTR pszCaption, UINT nStyle);
+int wbMessageBox(PWBOBJ pwboParent, LPCTSTR pszText, LPCTSTR pszCaption, UINT64 nStyle);
 BOOL wbExec(LPCTSTR pszPgm, LPCTSTR pszParm, BOOL bShowWindow);
-BOOL wbFindFile(LPTSTR pszFileName, UINT uLen);
+BOOL wbFindFile(LPTSTR pszFileName, UINT64 uLen);
 BOOL wbReadRegistryKey(LPCTSTR pszKey, LPTSTR pszSubKey, LPTSTR pszEntry, LPTSTR pszValue, int *pnBufLen);
 BOOL wbWriteRegistryKey(LPCTSTR pszKey, LPTSTR pszSubKey, LPTSTR pszEntry, LPCTSTR pszValue, DWORD dwValue, BOOL bString);
-LONG_PTR wbGetSystemInfo(LPCTSTR pszInfo, BOOL *pbString, LPTSTR pszString, UINT uLen);
+LONG_PTR wbGetSystemInfo(LPCTSTR pszInfo, BOOL *pbString, LPTSTR pszString, UINT64 uLen);
 BOOL wbSetAccelerators(PWBOBJ pwbo, LPACCEL paccels, int nCount);
 DWORD wbMakeAccelFromString(const char *pszAccel);
-UINT wbCheckInput(PWBOBJ pwbo, DWORD dwFlags, DWORD dwTimeout);
+UINT64 wbCheckInput(PWBOBJ pwbo, DWORD dwFlags, DWORD dwTimeout);
 BOOL wbSetCursor(PWBOBJ pwbo, LPCTSTR pszCursor, HANDLE handle);
 
 // Library-dependent functions
 
-UINT wbCallUserFunction(LPCTSTR pszFunctionName, LPDWORD pszObject, PWBOBJ pwboParent, PWBOBJ pctrl, UINT id, LPARAM lParam1, LPARAM lParam2, LPARAM lParam3);
+UINT64 wbCallUserFunction(LPCTSTR pszFunctionName, LPDWORD pszObject, PWBOBJ pwboParent, PWBOBJ pctrl, UINT64 id, LPARAM lParam1, LPARAM lParam2, LPARAM lParam3);
 BOOL wbError(LPCTSTR szFunction, int nType, LPCTSTR pszFmt, ...);
 void *wbMalloc(size_t size);
 void *wbCalloc(size_t nmemb, size_t size);
@@ -626,7 +626,7 @@ BOOL wbFree(void *ptr);
 void *wbRealloc(void *ptr, size_t size);
 char *wbStrDup(const char *string);
 char *wbStrnDup(const char *string, size_t size);
-UINT MemCheck(const char *message, BOOL mb);
+UINT64 MemCheck(const char *message, BOOL mb);
 
 //------------------------------------------------------------------ END OF FILE
